The biggest commit of all time ??? GONE WRONG.. hopefully not :)
I'd like to thank the academy for allowing me to commit something this big, truly was the work of a million men but in this case that million was literally just me.

When they ask about this commit, tell them Stang was one of the best to ever do it. He changed 56 files and added idk like 26,000 lines of code. NBD just another day for Stang...

Real commit notes: moved all the files out of source and into root. :)